requirements.txt
discord.py ~= 1.7.3
genshin ~= 0.4.0
python-dotenv ~= 0.20.0